chore(java): merge reflect.Types into TypeRef (#1731)
## What does this PR do?

This PR moves `org.apache.fury.reflect.Types` into TypeRef. Types is
used in Fury type system, by merge `org.apache.fury.reflect.Types`, we
can reduce ambiguation in fury type system

## Related issues
#1553 
#1690 

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fury/issues/new/choose) describing the
need to do so and update the document if necessary.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?


## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.
-->
