+++ java/fury-core/src/main/resources/META-INF/native-image/org.apache.fury/fury-core/native-image.properties
@@ -169,8 +169,8 @@ Args=--initialize-at-build-time=org.apache.fury.memory.MemoryBuffer,\
     org.apache.fury.meta.MetaStringEncoder$1,\
     org.apache.fury.meta.MetaString$Encoding,\
     org.apache.fury.meta.Encoders,\
-    org.apache.fury.reflect.Types$ClassOwnership,\
-    org.apache.fury.reflect.Types$ClassOwnership$1,\
-    org.apache.fury.reflect.Types$ClassOwnership$2,\
+    org.apache.fury.reflect.TypeRef$ClassOwnership,\
+    org.apache.fury.reflect.TypeRef$ClassOwnership$1,\
+    org.apache.fury.reflect.TypeRef$ClassOwnership$2,\
     org.apache.fury.resolver.DisallowedList,\
     org.apache.fury.util.StringUtils
